Basic implementation for adding words to local DB
Added predicates to 312-pess-grammar.pl for adding non-existing local
db words to the local db from the external Wordnet db. The predicates
makes use of the Morphological analysis tool morph_atoms_bag/2 to
produce stems of the given word and returns them in a list. The list of
stems is then check against our local DB to see if it exists, if not,
it will check the Wordnet DB instead and if it exists there, it will
add it to our local db.