Dictionary insertion now includes all word types
Before the code would only input the first word type returned by
Wordnet, however, some words are both nouns, and verbs or more. The new
code will now add all those if they do not exist in our library. If it
exists as a noun and not a verb it will add it as a verb. Also removed
a write predicate in stem_word(X).